backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-decouple==3.8
selenium==4.15.2
selenium-stealth==1.0.6
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
pandas==2.1.3
celery==5.3.4
redis==5.0.1
httpx==0.25.2
fake-useragent==1.4.0
python-dateutil==2.8.2

